requirements.txt
streamlit
numpy
Pillow
plotly
requests
openai
trimesh